Adds AAC to fips-value endpoint for cancer tables (#18)
This PR adds a field, `AAC`, present in the source cancer data that was
previously not being sent to the frontend. The PR includes changes to
the frontend to accommodate the change in the backends' schema, and
includes the new `AAC` value, if available, in the geometry popup
displayed by clicking on a county or tract.

**Specifics on the schema change:**
Previously, the `fips-value` endpoint would return a dict of FIPS
identifiers to numeric values; in the case of cancer data, this was the
`AAR` field. Now, the `fips-value` endpoint returns a dict of FIPS
identifiers to a dict of the form `{ v: number, aac?: number }`, where
the `v` subfield has taken the place of the numeric value previously
returned. For all non-cancer endpoints, `aac` is null.