Adds endpoint to download all data (#50)
This PR introduces a new API endpoint, `/stats/download-all`, that
returns a ZIP file to the user of all the measures served on the site.

Closes #8.

The ZIP file has the following structure:
- type (either "county" or "tract")
  - measure category (e.g., "Cancer Mortality (age-adj per 100k)")
    - measure CSVs (e.g., "Esophagus.csv")

The names for the measure category folders and measure files should
match what is displayed in the frontend. This probably isn't a big
concern, but we might consider either sanitizing them or changing to
their internal IDs if we anticipate that future CancerInFocus updates
will introduce characters that aren't allowed in filenames.

Internally, it iterates over all the individual CSV-downloading
endpoints and queries them individually, compiling the result into a ZIP
file which is then streamed to the user. All this is currently done in
memory, which should be fine considering the resulting uncompressed ZIP
file is ~10MB.

---------

Co-authored-by: Vincent Rubinetti <vince.rubinetti@gmail.com>