Simplify map no data, support explicit scale, add learn more links (#58)
Closes #54 
Closes #57
Related to #56 

- do data suppression (e.g., round cancer incidence/mortality down to 0
if small) at api level for simplicity. remove from values, so value
undefined and treated as "no data".
- add heading component for sources page. adds anchor links for easier
linking to sections.
- tweak link component
- map component now supports "explicit scale" that maps specific values
to specific colors
- map now displays "no data" legend and key entry automatically when any
of the map's regions are colored gray/undefined
- add "learn more about selected data" link next to map, which links to
specific sections on the sources page based on what is selected
- change behavior of factor selection re
https://github.com/colorado-cancer-center/ecco/pull/56#issuecomment-2079678800.
if selections change such that the selected factor value no longer
exists/is valid, instead of unsetting it, make it fallback to "All" or
the first available option. benefit: don't get "no selection" in
dropdowns due to invalid values for a factor. tradeoff: if user were to
go back to factor where value was valid, it won't be remembered.
- util funcs to scroll to hash

---------

Co-authored-by: Faisal Alquaddoomi <faisal.alquaddoomi@cuanschutz.edu>