Changing over to coe-ecco.org (#60)
This PR introduces a few backend changes to accommodate running the
frontent at coe-ecco.org via Netlify and the API server at
api.coe-ecco.org.

Specifically, the nginx configuration in `default.conf` has been split
into two server blocks:
- one that handles `api.coe-ecco.org` and redirects to the backend, and
- the other handles `coe-ecco.org` and serves the built frontend bundle.

I added `api.coe-ecco.org` and `coe-ecco.org` to the list of domains in
`DOMAIN_NAMES`. The result is that Let's Encrypt assigns the first value
in the list, e.g. `api.coe-ecco.org`, as the "canonical name" and uses
that domain as the name of the folder in which it stores keys. The other
domain names are registered as Subject Alternative Names on the same
certificate, meaning that certificate satisfies requests for those
domain names as well.

Because `coe-ecco.org` currently points to Netlify, we can't perform
HTTP-01 challenges for that domain; I manually performed a DNS-01
challenge for that domain for now, but if we're committed to Netlify I
could simply drop it from domain list, and I'd likely also remove
building and serving the frontend from the stack entirely.

I also added a few env vars related to the separate domains:
- `API_DOMAIN` is used by nginx to determine the hostname to direct to
the backend
- `FRONTEND_DOMAIN` is used by nginx the same way, but for the frontend
- `PRIMARY_DOMAIN` specifies the first entry in `DOMAIN_NAMES`
- it shouldn't be necessary to specify it here since the nginx
container's entrypoint script should set it as an env var by extracting
it from `DOMAIN_NAMES`, but for some reason it wasn't propagating into
nginx's template substitution layer, so I just pulled it out as an env
var for now.

---------

Co-authored-by: Vincent Rubinetti <vince.rubinetti@gmail.com>