Adds units to API response (#67)
This PR alters the `fips-value` endpoint to include a field named
`unit`, which is either `null` or is one of the following string values:

- **"percent"**: a percentage; the datasets we use encode these as
floats between 0 and 1, inclusive
- **"count"**: a raw count, represented as an integer, e.g. number of
cases
- **"rate"**: a rate of occurrence, represented as a float and typically
normalized by population
- **"dollar_amount"**: raw dollar amount, e.g. income
- **"rank"**: a position in a ranking, e.g. for the cancer disparities
index ranking of CO counties
- **"ordinal"**: a discrete value that can be ordered, e.g. "rising"
(1), "falling" (2), "stable" (3)
- **"categorical"**: a discrete value that *can't* be ordered, e.g.
"white", "black", "hispanic"
- **"ratio"**: range from -1 to +1, where -1 is "least affected" and +1
is "most affected"

Some of these units aren't currently in use, e.g. "categorical", and I
suspect many of them can be handled in the same way in the frontend.

The PR includes some simple tests that the `unit` field is included in
the `fips-value` response and that it matches what's defined in the
backend metadata.

The PR also includes a helper class, `MeasureMapper`, a dict-like that
when queried for a key produces a dict of the form `{label: <key>, unit:
<default_unit>}` for measures where the ID and the label are the same,
and all the measures have the same unit. A test is included for it as
well.

---------

Co-authored-by: Vincent Rubinetti <vince.rubinetti@gmail.com>