Automated deployment to backend VM (#68)
This PR introduces a new GitHub worflow, `deploy-prod.yaml`, that on a
merge into `main` does the following:

1. SSHes into the production VM `api.coe-ecco.org` as `ecco-deployer`;
the private SSH key is retrieved from the GitHub repo's secrets
2. Enters the deployment folder, `/var/projects/ecco`, on the VM
3. Pulls the latest changes to `main`
4. Appends the current commit hash and system date to `deploys.txt`
5. Launches `./run_stack.sh prod`, which rebuilds images and launches
the production version of the stack
6. Waits on the frontend container to finish building the site, then
exits with its exit status

I've had trouble getting this to work with
[act](https://github.com/nektos/act); sometimes it works, sometimes it
fails inexplicably. That's not much of a surprise for `act`, though;
it's often flakey, even if the workflow would work when deployed. I am
planning to continue to resolve the issues with `act` as it would be a
nice way to test workflows prior to merging, but it's not a big priority
IMHO. I'm hoping to debug this workflow, if necessary, on GitHub itself.