Loads database dump on every startup (#85)
*TL;DR: this PR loads dumps on every startup, not just when the database
is empty, removing the need to first delete the current database via
`down -v` and then restart the stack to reload a dump when the database
changes.*

This PR patches the postgis image's init script so that it executes
scripts located in `/docker-entrypoint-postinit.d/` every time the
database container starts. It operates in a similar fashion to how the
image executes scripts from `/docker-entrypoint-initdb.d/`, but unlike
those scripts the post-init ones are executed on every container
startup.

This PR also moves `load_latest_dump.sh` into the post-init scripts
folder, so that it loads the latest dump on every startup, not just when
there's no database. This has the effect of destroying any changes you
might have made to the database while it was running, so be sure to
enter the `db` container and export the database via the
`/db-exports/make_db_export.sh` script if there's anything you want to
save for next time.

While this scheme works for now, the behavior should probably be
revisited for a few reasons:
- while the database dump is currently small enough that it can be
quickly loaded on each execution, it might become expensive in the
future
- any data that's written to the database is essentially ephemeral,
which would become a problem if we start using it for, e.g., user logins
or user-submitted data
- if we upgrade to a new version of the postgres/postgis image and the
init script changes, we'll have to reimplement the patch