Fixes the by-county endpoint, other testing + module refactors (#99)
This PR addresses issues that lead to PR #93 having to be rolled back.
Specifically, the query that was being created to constrain factor
values to just values that actually occur in the dataset had a few
issues:
- the double-for loop comprehension I was using had some kind of issue
that went away when I switched to a less-clever nested for loop
- part of it was indented incorrectly, leading to multiple redundant
clauses being added to the query

The PR also factors out the factor resolution code into a new "helper"
module, `queries.py`. We may use the code in other parts of the stack,
e.g. in either the `fips-value` endpoint code to resolve factor values
when none have been supplied (currently it just uses the default
values), or possibly in the `measures` endpoint to dynamically provide
more useful factor value defaults than what we've hardcoded.

The PR adds the ability for `MeasureMapper` to optionally provide a list
of measures observed in the data when given a model. The change was
mostly to support testing SCP ("State Cancer Profiles") models: we
iterate over all of the measure categories and measures using the
metadata, and since the metadata for SCP is dynamically created (unlike
for CancerInFocus, where measure is hardcoded in the metadata) this was
previously impossible. Since we can't be sure `MeasureMapper` will be
used from an async context, I've added support functions to `db.py` to
get a synchronous session, which is used in `MeasureMapper`.

Finally, the PR includes a few new tests and organizes the existing ones
into subfolders. I added a README to help explain the organization
scheme. There are a few other minor changes to tests, e.g. updating
pytest annotations and skipping tests that can't be completed in
production due to the ingest data not being present.

**Things to try:**
- bring up the stack via `./run_stack.sh` and see that it completes
- enter the backend container via `./run_stack.sh shell` and run
`pytest` to see that the tests complete (they should take ~3 minutes)
- while a test exists to compare by-county to fips-value values, feel
free to spot-check the values shown via tooltips on the map against the
ones in the by-county view