Add health stats regions, youth vaping data (#106)
This PR introduces the health statistic regions (HSRs) requested in
#105. The source data for the HSRs was downloaded as GeoJSON from CDPHE
ArcGIS instance, i.e. [CDPHE Colorado Health Statistics
Regions](https://data-cdphe.opendata.arcgis.com/datasets/75e32548d3b24169adb942ecb7424937_6/explore?location=38.827448%2C-105.550900%2C7.05&showTable=true).

While we don't yet have the youth vaping data that will be associated
with the HSRs, I've done a first pass at modeling what the data might
look like and added a stand-in loading command,
`./commands/import_vaping_data.py`, which fabricates dummy data for the
`value` column over the product of all the age, race, and gender values.
The values for those factors were lifted from factor values we use in
other models for each HSR, but we'll replace them with the actual values
from the data once we've obtained it.

This PR will remain a draft until we've received the youth vaping data.
At that point we'll rewrite the import command to load from the actual
data source, replace the dummy data with actual data in the db dump, and
then un-draft the PR for final review.

**Schema note:** While HSRs are identified by the `hs_region` column in
the original data, I've duplicated this column and named it `FIPS` in
the imported table. This helps it match the schema of the existing
`County` and `Tract` models and allows it to be used in places where we
use those models. Since it's not really a FIPS code, I might walk that
back and implement a more flexible way to identify what the unique ID is
for a geometric entity, versus relying on a column named "FIPS" to
indicate that.

Some things to test:
- in the frontend, see that the "Health Statistics Regions" geometry
type shows up in "Geographic Level" dropdown
- see that "Youth Vaping" shows up as the only measure category
- see that "Youth Vaping Rate" shows up as the measure within that
category
- adjust the factors for that measure and see that the values change on
the map

Closes #105.

---------

Co-authored-by: Vincent Rubinetti <vince.rubinetti@gmail.com>