Alternate compare UI (#138)
Response to
https://github.com/colorado-cancer-center/ecco/pull/133#pullrequestreview-3281952292

> it's unclear to me that something has changed when you add the first
map.

Hopefully this more obvious now with the plus sign, map thumbnail, and
animations

> I would have expected the new measure to add to be near the "Add"
button, rather than where you make the main selection for what the map
displays.

> You _could_ add a category/measure selection right above the "Add"
button so that it doesn't change the page state until the user adds the
new comparison

The "main selection area" takes up quite a bit of space; 4 separate
dropdowns. Even if we only needed one extra set of dropdowns, it'd take
up too much space, given that the group has had lots of comments on
efficient use of space.

Side note: I don't want to hard-code anything into the design, like only
allowing 2 maps. Since I'm making the effort to refactor/rearrange
things now, I don't want to have to do it again when chance someone will
probably want > 2 maps in the future.

> When the preview is disabled it's IMO very confusing to no longer see
anything change when you change the measure.

> I also think "Show Preview" should just be enabled rather than
toggleable, since when it's disabled the UI can seem unresponsive if the
user isn't paying attention.

I added this for a specific scenario. It may be a rare scenario, but I
think it'd be frustrating for someone in it. Example:

You add 3 maps to the compare group. Then you go to a new measure,
thinking you might want to add it to the comparison. Then you reconsider
and don't want it. Then you want to display the first 3 maps in full
screen, because you're giving a presentation. But the 4th "tentative"
gray map is still on the screen. The only way to make it go away is to
change the dropdown selections back to one of the 3 maps in the
comparison group.

You can see this situation if you play around with it and imagine the
checkbox isn't there.

In this PR, I changed the checkbox to a one-way button (hide preview,
`showPreview = false`), made it only appear when there is a preview
(grayed-out tentative map) being shown, and made it reset/default to
`true` when there is any change to avoid accidentally leaving it on
`false`. I'm not really sure this makes it less confusing vs. a
checkbox, but at least it more narrowly targets the scenario above.

> Apparently it becomes "Remove" if the last map you added matches the
currently-selected measure, but I didn't immediately understand that it
was context-sensitive.

To be more precise, it's "remove" if the currently selected map
(measure/locations/etc) is in the compare group (doesn't have to be
last) and "add" if it's not.

Hopefully the changing of the tooltip text in this PR makes this more
clear. I don't know how else I would indicate it.

> IMO, it would be easier to understand if, rather than showing just the
number of maps in the comparison, you instead showed a list of the maps
that were currently in the comparison set with an option in each item to
remove that map.

I explored doing this in the previous PR as a list of the selected
level/category/measure/etc., but found that it was too verbose and just
duplicating text that was already on the screen (in the map top left
legends).

In this PR, we're still duplicating a lot of graphics with the map
thumbnails and they're not useful to look at individually due to being
small, but at least they're prettier, it's more visual than just a #
count, and hopefully it's more clear that we're comparing _full maps_.

Side note: we can't refer to this as "selected measures" because it's
also geographic levels, categories, factors, locations, and possibly
different scale colorings (I predict someone will request this). In this
PR I tried to make it more clear that "selected map" means "what's
selected in the various dropdowns/controls".