Fixes endpoint caching, makes more endpoints cached (#153)
While the current ECCO stack includes components for caching requests,
we were getting cache misses every time due to the fact that the library
we're using, https://github.com/long2ice/fastapi-cache, uses hashed
representations of the arguments to each function as the cache key. We
pass a database handle to most endpoints that's recreated on each
request, which means a new request can never match the cache.

Instead, [per the advice of the caching library
docs](https://github.com/long2ice/fastapi-cache?tab=readme-ov-file#custom-key-builder),
I've changed the key builder to one that uses just the request method,
path, and query parameters to construct the cache key. This leads to
demonstrably faster retrieval times for the cached endpoints --
`/locations`, for example, went from ~7 seconds to ~70ms.

This PR adds the `@cache` decorator to the geometry and locations
endpoints, and should fix the existing cache decorations on the
`/measures` and `/by-county/<FIPS>` endpoints.

On cache eviction/staleness: The cache we use is an in-memory cache held
in `memcached`, so it's evicted whenever the stack is restarted. This
occurs whenever we do a redeploy, usually triggered by a merge into
`main`. Since the database is read-only, the response would only change
if there were code and/or database changes, both of which can only occur
after the stack is restarted.